+++ src/EncompassRest/DirtyExtensibleObject.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Linq;
+using System.Reflection;
 using System.Runtime.CompilerServices;
 using EncompassRest.Utilities;
 using Newtonsoft.Json.Serialization;
@@ -119,11 +121,23 @@ namespace EncompassRest
             }
         }
         bool IDirty.Dirty { get => Dirty; set => Dirty = value; }
+
+        internal string LastId { get; set; }
+
         string IIdentifiable.Id { get => string.Empty; set { } }
 
         internal DirtyExtensibleObject()
         {
         }
+
+        internal static string GetIdPropertyName(TypeInfo typeInfo)
+        {
+            var idProperty = GetIdProperty(typeInfo);
+            var idPropertyNameAttribute = idProperty.GetCustomAttribute<IdPropertyNameAttribute>(false);
+            return idPropertyNameAttribute != null ? idPropertyNameAttribute.IdPropertyName : "Id";
+        }
+
+        private static PropertyInfo GetIdProperty(TypeInfo typeInfo) => typeInfo.DeclaredProperties.FirstOrDefault(p => p.Name == "EncompassRest.IIdentifiable.Id") ?? typeInfo.DeclaredProperties.FirstOrDefault(p => p.Name == "Id") ?? GetIdProperty(typeInfo.BaseType.GetTypeInfo());
 
 #if HAVE_ICLONEABLE
         object ICloneable.Clone() => this.Clone();
