+++ src/EncompassRest/ExtensibleObject.cs
@@ -16,23 +16,8 @@ namespace EncompassRest
         [JsonExtensionData]
         public IDictionary<string, object> ExtensionData { get => _extensionData ?? (_extensionData = new DirtyDictionary<string, object>(StringComparer.OrdinalIgnoreCase)); set => _extensionData = new DirtyDictionary<string, object>(value, StringComparer.OrdinalIgnoreCase); }
 
-        internal string LastId { get; set; }
-
         internal ExtensibleObject()
         {
         }
-
-#if HAVE_ICLONEABLE
-        object ICloneable.Clone() => this.Clone();
-#endif
-
-        internal static string GetIdPropertyName(TypeInfo typeInfo)
-        {
-            var idProperty = GetIdProperty(typeInfo);
-            var idPropertyNameAttribute = idProperty.GetCustomAttribute<IdPropertyNameAttribute>(false);
-            return idPropertyNameAttribute != null ? idPropertyNameAttribute.IdPropertyName : "Id";
-        }
-
-        private static PropertyInfo GetIdProperty(TypeInfo typeInfo) => typeInfo.DeclaredProperties.FirstOrDefault(p => p.Name == "EncompassRest.IIdentifiable.Id") ?? typeInfo.DeclaredProperties.FirstOrDefault(p => p.Name == "Id") ?? GetIdProperty(typeInfo.BaseType.GetTypeInfo());
     }
 }
